# backend/openapi.json
{
  "components": {
    "schemas": {
      "AIProviderStatus": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "available"
        ],
        "title": "AIProviderStatus",
        "type": "object"
      },
...
      "ApplyRulesRequest": {
        "additionalProperties": false,
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AssignmentRule"
            },
            "minItems": 1,
            "title": "Rules",
            "type": "array"
          }
        },
        "required": [
          "rules"
        ],
        "title": "ApplyRulesRequest",
        "type": "object"
      },
      "ApplySuggestionsRequest": {
        "additionalProperties": false,
        "description": "User decisions on a previously returned suggestion batch.",
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/SuggestionDecision"
            },
            "minItems": 1,
            "title": "Decisions",
            "type": "array"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "default": "unknown",
            "title": "Provider",
            "type": "string"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/MaterialSuggestion"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "decisions"
        ],
        "title": "ApplySuggestionsRequest",
        "type": "object"
      },
...
      "AssignResponse": {
        "additionalProperties": false,
        "properties": {
          "skipped_prim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Prim Ids",
            "type": "array"
          },
          "updated_prim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Updated Prim Ids",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "updated_prim_ids"
        ],
        "title": "AssignResponse",
        "type": "object"
      },
      "AssignmentRule": {
        "additionalProperties": false,
        "description": "One name-match -> RF material rule (SEAM spec style).\n\nA rule fires when any of ``match_name_contains`` appears (case-insensitively)\nin a prim's evidence (name / mesh_name / semantic tags / visual material\nname). ``rf_material_id`` must exist in the project library - unknown ids\nare dropped at generation time, same anti-hallucination stance as the\nsuggestion parser.",
        "properties": {
          "id": {
            "pattern": "^[a-z0-9_\\-]+$",
            "title": "Id",
            "type": "string"
          },
          "match_name_contains": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Match Name Contains",
            "type": "array"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "rf_material_id": {
            "title": "Rf Material Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "match_name_contains",
          "rf_material_id"
        ],
        "title": "AssignmentRule",
        "type": "object"
      },
...
      "ExplainValidationResponse": {
        "additionalProperties": false,
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "explanation",
          "provider"
        ],
        "title": "ExplainValidationResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaterialAlternative": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "rf_material_id": {
            "title": "Rf Material Id",
            "type": "string"
          }
        },
        "required": [
          "rf_material_id",
          "confidence"
        ],
        "title": "MaterialAlternative",
        "type": "object"
      },
...
      "MaterialSuggestion": {
        "additionalProperties": false,
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/MaterialAlternative"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "needs_user_confirmation": {
            "default": true,
            "title": "Needs User Confirmation",
            "type": "boolean"
          },
          "prim_id": {
            "title": "Prim Id",
            "type": "string"
          },
          "recommended_rf_material_id": {
            "title": "Recommended Rf Material Id",
            "type": "string"
          }
        },
        "required": [
          "prim_id",
          "recommended_rf_material_id",
          "confidence"
        ],
        "title": "MaterialSuggestion",
        "type": "object"
      },
      "MaterialSuggestionResponse": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/MaterialSuggestion"
            },
            "title": "Suggestions",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "provider"
        ],
        "title": "MaterialSuggestionResponse",
        "type": "object"
      },
...
      "RuleGenerationRequest": {
        "additionalProperties": false,
        "properties": {
          "instruction": {
            "minLength": 1,
            "title": "Instruction",
            "type": "string"
          }
        },
        "required": [
          "instruction"
        ],
        "title": "RuleGenerationRequest",
        "type": "object"
      },
      "RuleGenerationResponse": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AssignmentRule"
            },
            "title": "Rules",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "provider"
        ],
        "title": "RuleGenerationResponse",
        "type": "object"
      },
...
      "SuggestMaterialsRequest": {
        "additionalProperties": false,
        "properties": {
          "attach_texture_crops": {
            "default": false,
            "title": "Attach Texture Crops",
            "type": "boolean"
          },
          "prim_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prim Ids"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "screenshot_data_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screenshot Data Url"
          },
          "screenshot_data_urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 6,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screenshot Data Urls"
          }
        },
        "title": "SuggestMaterialsRequest",
        "type": "object"
      },
      "SuggestionDecision": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "approve",
              "reject",
              "edit"
            ],
            "title": "Action",
            "type": "string"
          },
          "prim_id": {
            "title": "Prim Id",
            "type": "string"
          },
          "rf_material_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rf Material Id"
          }
        },
        "required": [
          "prim_id",
          "action"
        ],
        "title": "SuggestionDecision",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Unified RF-visual scene authoring, RF material assignment, Sionna RT projection compilation, and simulation result APIs.",
    "title": "SionnaTwin Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/ai/apply-rules": {
      "post": {
        "operationId": "apply_rules_api_projects__project_id__ai_apply_rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRulesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialSuggestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Rules",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/projects/{project_id}/ai/apply-suggestions": {
      "post": {
        "operationId": "apply_suggestions_api_projects__project_id__ai_apply_suggestions_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplySuggestionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Suggestions",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/projects/{project_id}/ai/explain-validation": {
      "post": {
        "operationId": "explain_validation_api_projects__project_id__ai_explain_validation_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain Validation",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/projects/{project_id}/ai/generate-rules": {
      "post": {
        "operationId": "generate_rules_api_projects__project_id__ai_generate_rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Rules",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/projects/{project_id}/ai/status": {
      "get": {
        "operationId": "ai_status_api_projects__project_id__ai_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AIProviderStatus"
                  },
                  "title": "Response Ai Status Api Projects  Project Id  Ai Status Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Status",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/projects/{project_id}/ai/suggest-materials": {
      "post": {
        "operationId": "suggest_materials_api_projects__project_id__ai_suggest_materials_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestMaterialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialSuggestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest Materials",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}